requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.25
django-celery-results==2.5.1
django-cors-headers==4.1.0
djangorestframework==3.15.1
importlib-metadata==4.13.0
kombu==5.2.4
prompt_toolkit==3.0.48
psycopg2==2.9.9
pytz==2024.2
redis==5.0.8
sqlparse==0.4.4
typing_extensions==4.7.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
